Move TextNormalisation into the transformer common lib
