+++ pipeline/transformer/transformer_common/src/main/scala/weco/pipeline/transformer/transformers/ParsedPeriod.scala
@@ -1,9 +1,9 @@
 package weco.pipeline.transformer.transformers
 
 import weco.catalogue.internal_model.identifiers.IdState
-import weco.catalogue.internal_model.text.TextNormalisation._
 import weco.catalogue.internal_model.work.Period
 import weco.pipeline.transformer.parse.PeriodParser
+import weco.pipeline.transformer.text.TextNormalisation._
 
 object ParsedPeriod {
   def apply(label: String): Period[IdState.Unidentifiable.type] = {
