+++ pipeline/transformer/transformer_miro/src/main/scala/weco/pipeline/transformer/miro/transformers/MiroSubjects.scala
@@ -1,9 +1,9 @@
 package weco.pipeline.transformer.miro.transformers
 
-import weco.catalogue.internal_model.text.TextNormalisation._
 import weco.catalogue.internal_model.identifiers.IdState
 import weco.catalogue.internal_model.work.{Concept, Subject}
 import weco.pipeline.transformer.miro.source.MiroRecord
+import weco.pipeline.transformer.text.TextNormalisation._
 
 trait MiroSubjects {
